feat!(tree): move the tree module to the private API
BREAKING CHANGE: this breaks user code that might be relying on
the lv_tree module. The migration path for these users is to use
the private API by either including "lvgl_private.h" or enabling
LV_USE_PRIVATE_API
