+++ include/lvgl/lvgl.h
@@ -152,7 +152,6 @@
 #include "misc/lv_async.h"
 #include "misc/lv_math.h"
 #include "misc/lv_ll.h"
-#include "misc/lv_tree.h"
 #include "osal/lv_os.h"
 #include "others/file_explorer/lv_file_explorer.h"
 #include "others/fragment/lv_fragment.h"
